Now reading/writing new compartment model
git-svn-id: https://svn.ebi.ac.uk/steinbeck-research/CheMet/trunk@1403 47e25a5e-c6eb-4c94-8020-4a4a81ee2334

+++ CheMet/CheMet-Core/src/main/java/uk/ac/ebi/chemet/entities/reaction/participant/AtomContainerParticipant.java
@@ -304,8 +304,8 @@ public class AtomContainerParticipant extends ParticipantImplementation<IAtomCon
         if (o instanceof CompartmentalisedParticipant) {
             CompartmentalisedParticipant<IAtomContainer, Double, Compartment> co = (CompartmentalisedParticipant<IAtomContainer, Double, Compartment>) o;
             if (this.compartment != null && co.getCompartment() != null) {
-                Integer c1rank = this.compartment.getRanking();
-                Integer c2rank = co.getCompartment().getRanking();
+                Byte c1rank = this.compartment.getRanking();
+                Byte c2rank = co.getCompartment().getRanking();
                 int compComparison = c1rank.compareTo(c2rank);
                 if (compComparison != 0) {
                     return compComparison;
