+++ CheMet/interface/src/main/java/uk/ac/ebi/interfaces/reaction/Compartment.java
@@ -57,5 +57,5 @@ public interface Compartment {
     public Set<String> getSynonyms();
 
 
-    public int getRanking();
+    public byte getRanking();
 }
